Data.Bitcoin.PaymentChannel.Internal.Types

Documentation

dUST_LIMIT :: Word64

mIN_CHANNEL_SIZE :: Word64

pROTOCOL_VERSION :: Version

newtype UnsignedPaymentTx

Constructors

CUnsignedPaymentTx

Fields

unsignedTx :: Tx

type FinalTx = Tx

data PaymentChannelState

Shared state object used by both value sender and value receiver.

Constructors

CPaymentChannelState

Fields

pcsParameters :: ChannelParameters

pcsFundingTxInfo :: FundingTxInfo

pcsPaymentConfig :: PaymentTxConfig

pcsValueLeft :: BitcoinAmount

Value left to send

pcsPaymentSignature :: Maybe PaymentSignature

Signature over payment transaction of value pcsValueLeft unless no payment has been made yet

data ChannelParameters

Holds information about the channel

Constructors

CChannelParameters

Fields

cpSenderPubKey :: PubKey

cpReceiverPubKey :: PubKey

cpLockTime :: BitcoinLockTime

Channel expiration date/time

data FundingTxInfo

Holds information about the Bitcoin transaction used to fund the channel

Constructors

CFundingTxInfo

Fields

ftiHash :: TxHash

Hash of funding transaction.

ftiOutIndex :: Word32

Index/"vout" of funding TxOut

ftiOutValue :: BitcoinAmount

Value of funding TxOut (channel max value)

data PaymentTxConfig

Holds information about how to construct the payment transaction

Constructors

CPaymentTxConfig

Fields

ptcSenderChangeScript :: ByteString

Value sender change scriptPubKey in Bitcoin payment transaction

ptcReceiverChangeScript :: ByteString

Value receiver destination scriptPubKey in Bitcoin payment transaction

data Payment

Used to transfer value from sender to receiver.

Constructors

CPayment

Fields

cpChannelValueLeft :: BitcoinAmount

Channel value remaining (pcsValueLeft of the state from which this payment was made)

cpSignature :: PaymentSignature

Payment signature

data PaymentSignature

Contains payment signature plus sig hash flag byte

Constructors

CPaymentSignature

Fields

psSig :: Signature

psSigHash :: SigHash