Data.Bitcoin.PaymentChannel.Util

Description

Utility functions for Data.Bitcoin.PaymentChannel.

Documentation

getFundingAddress :: ChannelParameters -> Address

Derive a Bitcoin address, for funding a payment channel, from ChannelParameters. The transaction which pays to this address is the channel funding transaction, and information about this transaction is contained in FundingTxInfo.

data BitcoinLockTime

Data type representing a Bitcoin LockTime, which specifies a point in time. Derive a BitcoinLockTime from a UTCTime using fromDate.

parseBitcoinLocktime :: Word32 -> BitcoinLockTime

Convert from Bitcoin format (Word32)

toWord32 :: BitcoinLockTime -> Word32

Convert to Bitcoin format (Word32)

fromDate :: UTCTime -> BitcoinLockTime

Convert a UTCTime to a BitcoinLockTime